requirements.txt
ipykernel
pymongo
pandas
numpy
matplotlib
seaborn
pyarrow
fastparquet
sklearn
scikit-learn
scipy
xgboost
imbalanced-learn
catboost
python-dotenv==0.21.0
boto3==1.24.76
s3fs
PyYAML
-e .